refactor(html-comments): rename to no-html-comments
