+++ docs/rules.md
@@ -5,7 +5,7 @@
 * [block-indentation](rule/block-indentation.md)
 * [eol-last](rule/eol-last.md)
 * [invocable-blacklist](rule/invocable-blacklist.md)
-* [html-comments](rule/html-comments.md)
+* [no-html-comments](rule/no-html-comments.md)
 * [img-alt-attributes](rule/img-alt-attributes.md)
 * [inline-link-to](rule/inline-link-to.md)
 * [no-inline-styles](rule/no-inline-styles.md)
