+++ lib/config/recommended.js
@@ -5,7 +5,7 @@ module.exports = {
     'block-indentation': 2,
     'no-debugger': true,
     'no-log': true,
-    'html-comments': true,
+    'no-html-comments': true,
     'nested-interactive': true,
     'self-closing-void-elements': true,
     'triple-curlies': true,
