+++ lib/rules/index.js
@@ -9,7 +9,7 @@ module.exports = {
   'no-input-tagname': require('./lint-no-input-tagname'),
   'no-unbound': require('./lint-no-unbound'),
   'no-outlet-outside-routes': require('./lint-no-outlet-outside-routes'),
-  'html-comments': require('./lint-html-comments'),
+  'no-html-comments': require('./lint-no-html-comments'),
   'img-alt-attributes': require('./lint-img-alt-attributes'),
   'no-inline-styles': require('./lint-no-inline-styles'),
   'link-rel-noopener': require('./lint-link-rel-noopener'),
