+++ lib/rules/lint-no-html-comments.js
